+++ gradle.properties
@@ -3,5 +3,5 @@ fiatVersion=1.11.0
 enablePublishing=false
 korkVersion=7.3.0
 spinnakerGradleVersion=7.0.1
-keikoVersion=3.0.0
+keikoVersion=3.1.0
 org.gradle.parallel=true
